refactor(app): better folder structure
