--- src/modules/Header/index.tsx
@@ -1,1 +1,0 @@
-export * from './Header';
