--- src/modules/LoginForm/api/index.tsx
@@ -1,1 +1,0 @@
-export * from './loginUser';
