--- src/modules/RegisterForm/api/index.tsx
@@ -1,1 +1,0 @@
-export * from './registerUser';
