--- src/modules/RegisterForm/index.tsx
@@ -1,1 +1,0 @@
-export * from './RegisterForm';
