+++ src/pages/LoginPage.tsx
@@ -1,0 +1,9 @@
+import { Login } from '@/modules';
+
+export const LoginPage = () => {
+  return (
+    <>
+      <Login />;
+    </>
+  );
+};
