--- src/pages/LoginPage/LoginPage.tsx
@@ -1,9 +1,0 @@
-import { LoginForm } from '@/modules/LoginForm';
-
-export const LoginPage = () => {
-  return (
-    <>
-      <LoginForm />;
-    </>
-  );
-};
