--- src/pages/LoginPage/index.tsx
@@ -1,1 +1,0 @@
-export * from './LoginPage';
