+++ src/pages/RegisterPage.tsx
@@ -1,0 +1,9 @@
+import { Register } from '@/modules';
+
+export const RegisterPage = () => {
+  return (
+    <>
+      <Register />;
+    </>
+  );
+};
