--- src/pages/RegisterPage/RegisterPage.tsx
@@ -1,9 +1,0 @@
-import { RegisterForm } from '@/modules/RegisterForm';
-
-export const RegisterPage = () => {
-  return (
-    <>
-      <RegisterForm />;
-    </>
-  );
-};
