--- src/pages/RegisterPage/index.tsx
@@ -1,1 +1,0 @@
-export * from './RegisterPage';
