+++ src/routes/__root.tsx
@@ -1,7 +1,7 @@
 import { createRootRoute, Outlet } from '@tanstack/react-router';
 import { TanStackRouterDevtools } from '@tanstack/router-devtools';
 
-import { Header } from '@/modules/Header';
+import { Header } from '@/modules';
 
 export const Route = createRootRoute({
   component: () => (
