+++ src/routes/auth/login.tsx
@@ -3,9 +3,5 @@ import { createFileRoute } from '@tanstack/react-router';
 import { LoginPage } from '@/pages/LoginPage';
 
 export const Route = createFileRoute('/auth/login')({
-  component: RouteComponent,
+  component: LoginPage,
 });
-
-function RouteComponent() {
-  return <LoginPage />;
-}
