+++ src/routes/auth/register.tsx
@@ -3,9 +3,5 @@ import { createFileRoute } from '@tanstack/react-router';
 import { RegisterPage } from '@/pages/RegisterPage';
 
 export const Route = createFileRoute('/auth/register')({
-  component: RouteComponent,
+  component: RegisterPage,
 });
-
-function RouteComponent() {
-  return <RegisterPage />;
-}
